Notice, Faq, Inquiry 이미지 업로드 수정 및 유효성검사 추가

+++ src/main/java/org/devkirby/hanimman/controller/InquiryController.java
@@ -4,11 +4,17 @@ import lombok.RequiredArgsConstructor;
 import org.devkirby.hanimman.dto.InquiryDTO;
 import org.devkirby.hanimman.dto.InquiryFileDTO;
 import org.devkirby.hanimman.dto.InquiryRequest;
+import org.devkirby.hanimman.entity.User;
 import org.devkirby.hanimman.service.InquiryService;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PageableDefault;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
+
+import java.io.IOException;
+import java.util.HashMap;
+import java.util.Map;
 
 @RestController
 @RequestMapping("/api/v1/inquiry")
@@ -17,8 +23,24 @@ public class InquiryController {
     private final InquiryService inquiryService;
 
     @PostMapping
-    public void createInquiry(@RequestBody InquiryRequest inquiryRequest) {
-        inquiryService.create(inquiryRequest.getInquiryDTO(), inquiryRequest.getInquiryFileDTO());
+    public Map<String, Object> createInquiry(@RequestBody InquiryDTO inquiryDTO, @AuthenticationPrincipal User loginUser) throws IOException {
+        Map<String, Object> map = new HashMap<>();
+        if(inquiryDTO.getTitle().length() > 255 || inquiryDTO.getTitle().isEmpty()) {
+            throw new IllegalArgumentException("제목의 길이는 1자 이상, 255자 이하여야 합니다. 현재 길이: "
+                    + inquiryDTO.getTitle().length());
+        }else if(inquiryDTO.getContent().length() > 65535) {
+            throw new IllegalArgumentException("내용의 길이는 65535자 이하여야 합니다. 현재 길이: "
+                    + inquiryDTO.getContent().length());
+        }else if(inquiryDTO.getFiles().size() > 10) {
+            throw new IllegalArgumentException("이미지는 최대 10개까지 업로드할 수 있습니다. 현재 이미지 개수: "
+                    + inquiryDTO.getFiles().size());
+        } else {
+            inquiryDTO.setUserId(loginUser.getId());
+            inquiryService.create(inquiryDTO);
+            map.put("code", 200);
+            map.put("msg", "1:1 문의 작성에 성공했습니다.");
+        }
+        return map;
     }
 
     @GetMapping("/{id}")
@@ -27,14 +49,38 @@ public class InquiryController {
     }
 
     @PutMapping("/{id}")
-    public void updateInquiry(@PathVariable Integer id, @RequestBody InquiryDTO inquiryDTO) {
+    public Map<String, Object> updateInquiry(@PathVariable Integer id, @RequestBody InquiryDTO inquiryDTO, @AuthenticationPrincipal User loginUser) throws IOException {
+        Map<String, Object> map = new HashMap<>();
         inquiryDTO.setId(id);
-        inquiryService.update(inquiryDTO);
+        if(inquiryDTO.getTitle().length() > 255 || inquiryDTO.getTitle().isEmpty()) {
+            throw new IllegalArgumentException("제목의 길이는 1자 이상, 255자 이하여야 합니다. 현재 길이: "
+                    + inquiryDTO.getTitle().length());
+        }else if(inquiryDTO.getContent().length() > 65535) {
+            throw new IllegalArgumentException("내용의 길이는 65535자 이하여야 합니다. 현재 길이: "
+                    + inquiryDTO.getContent().length());
+        }else if(inquiryDTO.getFiles().size() > 10) {
+            throw new IllegalArgumentException("이미지는 최대 10개까지 업로드할 수 있습니다. 현재 이미지 개수: "
+                    + inquiryDTO.getFiles().size());
+        } else {
+            inquiryDTO.setUserId(loginUser.getId());
+            inquiryService.create(inquiryDTO);
+            map.put("code", 200);
+            map.put("msg", "1:1 문의 수정에 성공했습니다.");
+        }
+        return map;
     }
 
     @DeleteMapping("/{id}")
-    public void deleteInquiry(@PathVariable Integer id) {
-        inquiryService.delete(id);
+    public Map<String, Object> deleteInquiry(@PathVariable Integer id, @AuthenticationPrincipal User loginUser) {
+        Map<String, Object> map = new HashMap<>();
+        if(!loginUser.getId().equals(inquiryService.read(id).getUserId())) {
+            throw new IllegalArgumentException("본인이 작성한 문의만 삭제할 수 있습니다.");
+        }else{
+            inquiryService.delete(id);
+            map.put("code", 200);
+            map.put("msg", "1:1 문의 삭제에 성공했습니다.");
+        }
+        return map;
     }
 
     @GetMapping
